requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
future==1.0.0
iso8601==2.1.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pillow==11.1.0
PyOpenGL==3.1.9
pyparsing==3.2.1
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
pyserial==3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.17.0
